+++ code/go/0chain.net/smartcontract/storagesc/alloation2_test.go
@@ -1,6 +1,13 @@
 package storagesc
 
 import (
+	"encoding/json"
+	"math"
+	"strconv"
+	"strings"
+	"testing"
+	"time"
+
 	cstate "0chain.net/chaincore/chain/state"
 	sci "0chain.net/chaincore/smartcontractinterface"
 	"0chain.net/chaincore/state"
@@ -9,13 +16,7 @@ import (
 	"0chain.net/core/common"
 	"0chain.net/core/datastore"
 	"0chain.net/core/util"
-	"encoding/json"
 	"github.com/stretchr/testify/require"
-	"math"
-	"strconv"
-	"strings"
-	"testing"
-	"time"
 )
 
 type blobberStakes []int64
@@ -111,7 +112,7 @@ func TestCancelAllocationRequest(t *testing.T) {
 	var challenges = [][]common.Timestamp{}
 	var scYaml = scConfig{
 		MaxMint: zcnToBalance(4000000.0),
-		StakePool: &stakePoolConfig{
+		StakePool: &StakePoolConfig{
 			InterestRate:     0.0000334,
 			InterestInterval: time.Minute,
 		},
@@ -265,7 +266,7 @@ func TestFinalizeAllocation(t *testing.T) {
 	var blobberStakePools = [][]mockStakePool{}
 	var scYaml = scConfig{
 		MaxMint: zcnToBalance(4000000.0),
-		StakePool: &stakePoolConfig{
+		StakePool: &StakePoolConfig{
 			InterestRate:     0.0000334,
 			InterestInterval: time.Minute,
 		},
