+++ code/go/0chain.net/smartcontract/storagesc/block_reward.go
@@ -1,3 +1,5 @@
+// +build depreciated
+
 package storagesc
 
 import (
