+++ code/go/0chain.net/smartcontract/storagesc/challenge_test.go
@@ -1,6 +1,14 @@
 package storagesc
 
 import (
+	"encoding/json"
+	"fmt"
+	"math/rand"
+	"strconv"
+	"strings"
+	"testing"
+	"time"
+
 	cstate "0chain.net/chaincore/chain/state"
 	sci "0chain.net/chaincore/smartcontractinterface"
 	"0chain.net/chaincore/state"
@@ -9,14 +17,7 @@ import (
 	"0chain.net/core/common"
 	"0chain.net/core/datastore"
 	"0chain.net/core/util"
-	"encoding/json"
-	"fmt"
 	"github.com/stretchr/testify/require"
-	"math/rand"
-	"strconv"
-	"strings"
-	"testing"
-	"time"
 )
 
 const (
@@ -256,7 +257,7 @@ func TestBlobberPenalty(t *testing.T) {
 	var otherWritePools = 4
 	var scYaml = scConfig{
 		MaxMint: zcnToBalance(4000000.0),
-		StakePool: &stakePoolConfig{
+		StakePool: &StakePoolConfig{
 			InterestRate:     0.0000334,
 			InterestInterval: time.Minute,
 		},
