+++ code/go/0chain.net/smartcontract/storagesc/config_settings_test.go
@@ -1,4 +1,4 @@
-package storagesc
+package storagesc_test
 
 import (
 	"fmt"
@@ -8,22 +8,24 @@ import (
 
 	"0chain.net/chaincore/state"
 
-	"0chain.net/smartcontract"
-
 	chainstate "0chain.net/chaincore/chain/state"
 	"0chain.net/chaincore/mocks"
 	sci "0chain.net/chaincore/smartcontractinterface"
 	"0chain.net/chaincore/transaction"
+	"0chain.net/smartcontract"
+	. "0chain.net/smartcontract/storagesc"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 )
+
+const owner = "1746b06bb09f55ee01b33b5e2e055d6cc7a900cb57c0a3a5eaabb8a0e7745802"
 
 func TestSettings(t *testing.T) {
 	require.Len(t, SettingName, int(NumberOfSettings))
 	require.Len(t, Settings, int(NumberOfSettings))
 
 	for _, name := range SettingName {
-		require.EqualValues(t, name, SettingName[Settings[name].setting])
+		require.EqualValues(t, name, SettingName[Settings[name].Setting])
 	}
 }
 
@@ -38,7 +40,7 @@ func TestUpdateConfig(t *testing.T) {
 	type parameters struct {
 		client       string
 		inputMap     map[string]string
-		TargetConfig scConfig
+		TargetConfig ScConfig
 	}
 
 	setExpectations := func(t *testing.T, p parameters) args {
@@ -50,15 +52,19 @@ func TestUpdateConfig(t *testing.T) {
 			ClientID: p.client,
 		}
 
-		balances.On("GetTrieNode", scConfigKey(ssc.ID)).Return(&scConfig{}, nil).Once()
+		balances.On("GetTrieNode", ScConfigKey(ssc.ID)).Return(&ScConfig{}, nil).Once()
 		balances.On(
 			"InsertTrieNode",
-			scConfigKey(ssc.ID),
-			mock.MatchedBy(func(conf *scConfig) bool {
+			ScConfigKey(ssc.ID),
+			mock.MatchedBy(func(in interface{}) bool {
+				conf, ok := IToScConfig(in)
+				if !ok {
+					return false
+				}
 				for key, value := range p.inputMap {
 					var setting interface{} = getConfField(*conf, key)
 					fmt.Println("setting", setting, "value", value)
-					switch Settings[key].configType {
+					switch Settings[key].ConfigType {
 					case smartcontract.Int:
 						{
 							expected, err := strconv.Atoi(value)
@@ -112,7 +118,7 @@ func TestUpdateConfig(t *testing.T) {
 					case smartcontract.StateBalance:
 						{
 							expected, err := strconv.ParseFloat(value, 64)
-							expected = x10 * expected
+							expected = 1e10 * expected
 							require.NoError(t, err)
 							actual, ok := setting.(state.Balance)
 							require.True(t, ok)
@@ -212,7 +218,7 @@ func TestUpdateConfig(t *testing.T) {
 			t.Parallel()
 			args := setExpectations(t, test.parameters)
 
-			_, err := args.ssc.updateSettings(args.txn, args.input, args.balances)
+			_, err := args.ssc.UpdateSettings(args.txn, args.input, args.balances)
 			require.EqualValues(t, test.want.error, err != nil)
 			if err != nil {
 				require.EqualValues(t, test.want.msg, err.Error())
@@ -223,8 +229,8 @@ func TestUpdateConfig(t *testing.T) {
 	}
 }
 
-func getConfField(conf scConfig, field string) interface{} {
-	switch Settings[field].setting {
+func getConfField(conf ScConfig, field string) interface{} {
+	switch Settings[field].Setting {
 	case MaxMint:
 		return conf.MaxMint
 	case TimeUnit:
