+++ code/go/0chain.net/smartcontract/storagesc/curator_test.go
@@ -1,14 +1,16 @@
-package storagesc
+package storagesc_test
 
 import (
+	"encoding/json"
+	"testing"
+
 	chainState "0chain.net/chaincore/chain/state"
 	"0chain.net/chaincore/mocks"
 	sci "0chain.net/chaincore/smartcontractinterface"
 	"0chain.net/chaincore/transaction"
-	"encoding/json"
+	. "0chain.net/smartcontract/storagesc"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
-	"testing"
 )
 
 func TestAddCurator(t *testing.T) {
@@ -26,7 +28,7 @@ func TestAddCurator(t *testing.T) {
 	}
 	type parameters struct {
 		clientId         string
-		info             curatorInput
+		info             CuratorInput
 		existingCurators []string
 	}
 	type want struct {
@@ -83,7 +85,7 @@ func TestAddCurator(t *testing.T) {
 			name: "ok",
 			parameters: parameters{
 				clientId: mockClientId,
-				info: curatorInput{
+				info: CuratorInput{
 					CuratorId:    mockCuratorId,
 					AllocationId: mockAllocationId,
 				},
@@ -96,7 +98,7 @@ func TestAddCurator(t *testing.T) {
 			t.Parallel()
 			args := setExpectations(t, test.name, test.parameters, test.want)
 
-			err := args.ssc.addCurator(args.txn, args.input, args.balances)
+			err := args.ssc.AddCurator(args.txn, args.input, args.balances)
 
 			require.EqualValues(t, test.want.err, err != nil)
 			if err != nil {
@@ -125,7 +127,7 @@ func TestRemoveCurator(t *testing.T) {
 	}
 	type parameters struct {
 		clientId         string
-		info             curatorInput
+		info             CuratorInput
 		existingCurators []string
 	}
 	type want struct {
@@ -179,7 +181,7 @@ func TestRemoveCurator(t *testing.T) {
 			name: "ok_1_curators",
 			parameters: parameters{
 				clientId: mockClientId,
-				info: curatorInput{
+				info: CuratorInput{
 					CuratorId:    mockCuratorId,
 					AllocationId: mockAllocationId,
 				},
@@ -190,7 +192,7 @@ func TestRemoveCurator(t *testing.T) {
 			name: "ok_3_curators",
 			parameters: parameters{
 				clientId: mockClientId,
-				info: curatorInput{
+				info: CuratorInput{
 					CuratorId:    mockCuratorId,
 					AllocationId: mockAllocationId,
 				},
@@ -201,7 +203,7 @@ func TestRemoveCurator(t *testing.T) {
 			name: "error_not_curator",
 			parameters: parameters{
 				clientId: mockClientId,
-				info: curatorInput{
+				info: CuratorInput{
 					CuratorId:    mockCuratorId,
 					AllocationId: mockAllocationId,
 				},
@@ -219,7 +221,7 @@ func TestRemoveCurator(t *testing.T) {
 			t.Parallel()
 			args := setExpectations(t, test.name, test.parameters, test.want)
 
-			err := args.ssc.removeCurator(args.txn, args.input, args.balances)
+			err := args.ssc.RemoveCurator(args.txn, args.input, args.balances)
 
 			require.EqualValues(t, test.want.err, err != nil)
 			if err != nil {
