+++ code/go/0chain.net/smartcontract/storagesc/functions_test.go
@@ -1,15 +1,17 @@
-package storagesc
+package storagesc_test
 
 import (
+	"strconv"
+	"testing"
+
 	cstate "0chain.net/chaincore/chain/state"
 	"0chain.net/chaincore/mocks"
 	"0chain.net/chaincore/state"
 	"0chain.net/chaincore/tokenpool"
 	"0chain.net/core/datastore"
+	. "0chain.net/smartcontract/storagesc"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
-	"strconv"
-	"testing"
 )
 
 func TestTransferReward(t *testing.T) {
@@ -24,7 +26,7 @@ func TestTransferReward(t *testing.T) {
 	type args struct {
 		sscKey   string
 		zcnPool  tokenpool.ZcnPool
-		sp       *stakePool
+		sp       *StakePool
 		value    state.Balance
 		balances cstate.StateContextI
 	}
@@ -56,18 +58,18 @@ func TestTransferReward(t *testing.T) {
 	var setup = func(t *testing.T, p parameters) args {
 		var zcnPool tokenpool.ZcnPool
 		zcnPool.Balance = zcnToBalance(2 * p.value)
-		sPool := newStakePool()
+		sPool := NewStakePool()
 
-		sPool.Settings = stakePoolSettings{
+		sPool.SetSettings(StakePoolSettings{
 			ServiceCharge:  p.serviceCharge,
 			DelegateWallet: p.to,
-		}
+		})
 		for i, d := range p.delegates {
 			id := strconv.Itoa(i)
-			dPool := delegatePool{}
+			dPool := DelegatePool{}
 			dPool.Balance = zcnToBalance(d)
 			dPool.DelegateID = id
-			sPool.Pools[id] = &dPool
+			sPool.SetPool(id, &dPool)
 		}
 
 		balances := setExpectations(t, p)
@@ -75,7 +77,7 @@ func TestTransferReward(t *testing.T) {
 		return args{
 			sscKey:   p.from,
 			zcnPool:  zcnPool,
-			sp:       sPool,
+			sp:       (*StakePool)(sPool),
 			value:    zcnToBalance(p.value),
 			balances: balances,
 		}
@@ -112,7 +114,7 @@ func TestTransferReward(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
 			args := setup(t, tt.parameters)
-			moved, err := transferReward(args.sscKey, args.zcnPool, args.sp, args.value, args.balances)
+			moved, err := TransferReward(args.sscKey, args.zcnPool, args.sp, args.value, args.balances)
 			require.EqualValues(t, tt.want.error, err != nil)
 			if err != nil {
 				require.EqualValues(t, tt.want.errorMsg, err.Error())
@@ -133,7 +135,7 @@ func TestMintReward(t *testing.T) {
 	}
 
 	type args struct {
-		sp       *stakePool
+		sp       *StakePool
 		value    float64
 		balances cstate.StateContextI
 	}
@@ -163,23 +165,23 @@ func TestMintReward(t *testing.T) {
 	}
 
 	var setup = func(t *testing.T, p parameters) args {
-		sPool := newStakePool()
-		sPool.Settings = stakePoolSettings{
+		sPool := NewStakePool()
+		sPool.SetSettings(StakePoolSettings{
 			ServiceCharge:  p.serviceCharge,
 			DelegateWallet: p.to,
-		}
+		})
 		for i, d := range p.delegates {
 			id := strconv.Itoa(i)
-			dPool := delegatePool{}
+			dPool := DelegatePool{}
 			dPool.Balance = zcnToBalance(d)
 			dPool.DelegateID = id
-			sPool.Pools[id] = &dPool
+			sPool.SetPool(id, &dPool)
 		}
 
 		balances := setExpectations(t, p)
 
 		return args{
-			sp:       sPool,
+			sp:       (*StakePool)(sPool),
 			value:    float64(zcnToBalance(p.value)),
 			balances: balances,
 		}
@@ -223,7 +225,7 @@ func TestMintReward(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
 			args := setup(t, tt.parameters)
-			err := mintReward(args.sp, args.value, args.balances)
+			err := MintReward(args.sp, args.value, args.balances)
 			require.EqualValues(t, tt.want.error, err != nil)
 			if err != nil {
 				require.EqualValues(t, tt.want.errorMsg, err.Error())
