+++ code/go/0chain.net/smartcontract/storagesc/funded_pools_test.go
@@ -1,13 +1,15 @@
-package storagesc
+package storagesc_test
 
 import (
+	"testing"
+
 	cstate "0chain.net/chaincore/chain/state"
 	"0chain.net/chaincore/mocks"
 	sci "0chain.net/chaincore/smartcontractinterface"
 	"0chain.net/core/util"
+	. "0chain.net/smartcontract/storagesc"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
-	"testing"
 )
 
 func TestAddToFundedPools(t *testing.T) {
@@ -29,22 +31,26 @@ func TestAddToFundedPools(t *testing.T) {
 			SmartContract: sci.NewSC(ADDRESS),
 		}
 		if len(p.existing) != 0 {
-			var existingPools fundedPools = p.existing
+			var existingPools FundedPools = p.existing
 			balances.On(
 				"GetTrieNode",
-				fundedPoolsKey(ssc.ID, p.client),
+				FundedPoolsKey(ssc.ID, p.client),
 			).Return(&existingPools, nil).Once()
 		} else {
 			balances.On(
 				"GetTrieNode",
-				fundedPoolsKey(ssc.ID, p.client),
+				FundedPoolsKey(ssc.ID, p.client),
 			).Return(nil, util.ErrValueNotPresent).Once()
 		}
 
 		balances.On(
 			"InsertTrieNode",
-			fundedPoolsKey(ssc.ID, p.client),
-			mock.MatchedBy(func(fp *fundedPools) bool {
+			FundedPoolsKey(ssc.ID, p.client),
+			mock.MatchedBy(func(ifp interface{}) bool {
+				fp, ok := IToFundedPool(ifp)
+				if !ok {
+					return false
+				}
 				if len(p.existing) != len(*fp)-1 {
 					return false
 				}
@@ -93,7 +99,7 @@ func TestAddToFundedPools(t *testing.T) {
 
 			ssc, balances := setExpectations(t, tt.parameters)
 
-			err := ssc.addToFundedPools(tt.parameters.client, tt.parameters.newPool, balances)
+			err := ssc.AddToFundedPools(tt.parameters.client, tt.parameters.newPool, balances)
 
 			require.EqualValues(t, tt.want.error, err != nil)
 			if err != nil {
@@ -125,15 +131,15 @@ func TestIsFundedPool(t *testing.T) {
 		}
 
 		if len(p.existing) != 0 {
-			var existingPools fundedPools = p.existing
+			var existingPools FundedPools = p.existing
 			balances.On(
 				"GetTrieNode",
-				fundedPoolsKey(ssc.ID, p.client),
+				FundedPoolsKey(ssc.ID, p.client),
 			).Return(&existingPools, nil).Once()
 		} else {
 			balances.On(
 				"GetTrieNode",
-				fundedPoolsKey(ssc.ID, p.client),
+				FundedPoolsKey(ssc.ID, p.client),
 			).Return(nil, util.ErrValueNotPresent).Once()
 		}
 
@@ -187,7 +193,7 @@ func TestIsFundedPool(t *testing.T) {
 
 			ssc, balances := setExpectations(t, tt.parameters)
 
-			ok, err := ssc.isFundedPool(tt.parameters.client, tt.parameters.poolId, balances)
+			ok, err := ssc.IsFundedPool(tt.parameters.client, tt.parameters.poolId, balances)
 
 			require.EqualValues(t, tt.want.error, err != nil)
 			if err != nil {
