+++ code/go/0chain.net/smartcontract/storagesc/sc.go
@@ -241,8 +241,6 @@ func (sc *StorageSmartContract) Execute(t *transaction.Transaction,
 		resp, err = sc.blobberHealthCheck(t, input, balances)
 	case "update_blobber_settings":
 		resp, err = sc.updateBlobberSettings(t, input, balances)
-	case "pay_blobber_block_rewards":
-		err = sc.payBlobberBlockRewards(balances)
 
 	// read_pool
